refactor: migrate subComponents to shared-components
