+++ docs/docs/auto-docs/components/EventCalender/Header/EventHeader/functions/default.md
@@ -6,7 +6,7 @@
 
 > **default**(`__namedParameters`): `Element`
 
-Defined in: [src/components/EventCalender/Header/EventHeader.tsx:47](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/EventCalender/Header/EventHeader.tsx#L47)
+Defined in: [src/components/EventCalender/Header/EventHeader.tsx:46](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/EventCalender/Header/EventHeader.tsx#L46)
 
 ## Parameters
