+++ docs/docs/auto-docs/shared-components/Navbar/Navbar/functions/default.md
@@ -6,7 +6,7 @@
 
 > **default**(`__namedParameters`): `Element`
 
-Defined in: [src/shared-components/Navbar/Navbar.tsx:86](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Navbar/Navbar.tsx#L86)
+Defined in: [src/shared-components/Navbar/Navbar.tsx:73](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Navbar/Navbar.tsx#L73)
 
 ## Parameters
