+++ docs/docs/auto-docs/shared-components/SearchingButton/SearchingButton/namespaces/default/variables/propTypes.md
@@ -1,0 +1,9 @@
+[Admin Docs](/)
+
+***
+
+# Variable: propTypes
+
+> **propTypes**: `any`
+
+Defined in: [src/shared-components/SearchingButton/SearchingButton.tsx:53](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/SearchingButton/SearchingButton.tsx#L53)
