+++ docs/docs/auto-docs/shared-components/SortingButton/SortingButton/namespaces/default/variables/propTypes.md
@@ -1,0 +1,9 @@
+[Admin Docs](/)
+
+***
+
+# Variable: propTypes
+
+> **propTypes**: `any`
+
+Defined in: [src/shared-components/SortingButton/SortingButton.tsx:108](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/SortingButton/SortingButton.tsx#L108)
