+++ docs/docs/auto-docs/shared-components/SortingButton/SortingButton/variables/default.md
@@ -6,7 +6,7 @@
 
 > `const` **default**: `React.FC`\<`InterfaceSortingButtonProps`\>
 
-Defined in: [src/subComponents/SortingButton.tsx:48](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/subComponents/SortingButton.tsx#L48)
+Defined in: [src/shared-components/SortingButton/SortingButton.tsx:48](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/SortingButton/SortingButton.tsx#L48)
 
 SortingButton component renders a Dropdown with sorting options.
 It allows users to select a sorting option and triggers a callback on selection.
