--- docs/docs/auto-docs/subComponents/SearchingButton/namespaces/default/variables/propTypes.md
@@ -1,9 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Variable: propTypes
-
-> **propTypes**: `any`
-
-Defined in: [src/subComponents/SearchingButton.tsx:53](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/subComponents/SearchingButton.tsx#L53)
