--- docs/docs/auto-docs/subComponents/SortingButton/namespaces/default/variables/propTypes.md
@@ -1,9 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Variable: propTypes
-
-> **propTypes**: `any`
-
-Defined in: [src/subComponents/SortingButton.tsx:108](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/subComponents/SortingButton.tsx#L108)
