+++ src/shared-components/PeopleTabNavbar/PeopleTabNavbar.tsx
@@ -7,7 +7,7 @@
  *
  * @remarks
  * - Used on pages that require filtering, sorting, or searching
- * - Uses SearchBar and SortingButton subcomponents
+ * - Uses SearchBar and SortingButton shared-components
  * - Layout adapts based on provided props
  *
  * @example
@@ -46,7 +46,7 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import styles from 'style/app-fixed.module.css';
 import SearchBar from 'shared-components/SearchBar/SearchBar';
-import SortingButton from 'subComponents/SortingButton';
+import SortingButton from 'shared-components/SortingButton/SortingButton';
 import { InterfacePeopleTabNavbarProps } from 'types/PeopleTab/interface';
 
 export default function PeopleTabNavbar({
