+++ src/shared-components/SearchFilterBar/SearchFilterBar.spec.tsx
@@ -38,7 +38,7 @@ vi.mock('shared-components/SearchBar/SearchBar', () => ({
   ),
 }));
 
-vi.mock('subComponents/SortingButton', () => ({
+vi.mock('shared-components/SortingButton/SortingButton', () => ({
   default: vi.fn(
     ({
       buttonLabel,
