+++ src/shared-components/SearchingButton/SearchingButton.spec.tsx
